FilterButtons as widget (#5547)
* FilterButtons as widget

* actually solve the on wiki usage
