+++ components/main_page/wikis/leagueoflegends/main_page_layout_data.lua
@@ -47,7 +47,7 @@ local CONTENT = {
 	filterButtons = {
 		noPanel = true,
 		body = '<div style{{=}}"width:100%;margin-bottom:8px;">' ..
-			'{{#invoke:Lua|invoke|module=FilterButtons|fn=getFromConfig}}</div>',
+			'{{#invoke:Lua|invoke|module=Widget/Factory|fn=fromTemplate|widget=FilterButtons}}</div>',
 	},
 	matches = {
 		heading = 'Matches',
